backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.1.0
python-dotenv==1.0.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
ultralytics==8.0.227
pandas==2.1.4
openpyxl==3.1.2
websockets==12.0
asyncpg==0.30.0
aiohttp==3.12.13
httpx==0.24.1